Add support for Spigot 1.10-1.12
